pyproject.toml
[project]
name = "screenchat-live"
version = "0.1.0"
description = "Terminal voice and screen-share chat client for Gemini Live"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.68.0",
    "mss>=10.1.0",
    "pillow>=12.1.1",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.2.2",
    "ruff>=0.15.8",
]

[tool.ruff]
line-length = 180

[tool.ruff.lint]
select = [
    "E4", "E7", "E9", "F",   # Pyflakes, pycodestyle
    "B", "B9",               # flake8-bugbear
    "I",                     # isort
    "UP",                    # pyupgrade
    "SIM",                   # flake8-simplify
    "C4",                    # flake8-comprehensions
    "ARG",                   # flake8-unused-arguments
    "DTZ",                   # flake8-datetimez
    "G",                     # flake8-logging-format
    "S",                     # flake8-bandit
    "RUF",                   # Ruff-specific rules
    "PT",                    # flake8-pytest-style
    "PIE",                   # flake8-pie
    "T20",                   # flake8-print
]
ignore = [
    "B008", "S101", "S105", "S106", "G004", "T201", "S112", "E731", "S311"
]
unfixable = ["DTZ", "S"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]
"tests/**/*" = ["S101", "ARG", "T20"]
"test_*.py" = ["S101", "ARG", "T20"]

[tool.ruff.lint.isort]
combine-as-imports = true
force-sort-within-sections = true

[tool.ruff.format]
docstring-code-format = true
